+++ scheduler/env.c
@@ -114,11 +114,14 @@ cupsdSetEnv(const char *name,		/* I - Name of variable */
     return;
 
  /*
-  * Do not allow dynamic linker variables when running as root...
+  * Do not allow dynamic linker variables when running as root and
+  * not being confined in a Snap...
   */
 
+#ifndef SUPPORT_SNAPPED_CUPSD
   if (!RunUser && (!strncmp(name, "DYLD_", 5) || !strncmp(name, "LD_", 3)))
     return;
+#endif /* !SUPPORT_SNAPPED_CUPSD */
 
  /*
   * See if this variable has already been defined...
@@ -208,9 +211,19 @@ cupsdUpdateEnv(void)
   set_if_undefined("LD_LIBRARY_PATH", NULL);
   set_if_undefined("LD_PRELOAD", NULL);
   set_if_undefined("NLSPATH", NULL);
+  /* Only if cupsd is confined in a Snap we pass the PATH environment
+     variable on for external executables we call */
   if (find_env("PATH") < 0)
-    cupsdSetEnvf("PATH", "%s/filter:" CUPS_BINDIR ":" CUPS_SBINDIR
-			 ":/bin:/usr/bin", ServerBin);
+  {
+#ifdef SUPPORT_SNAPPED_CUPSD
+    char *value;
+    if ((value = getenv("PATH")) != NULL)
+      cupsdSetEnvf("PATH", "%s/filter:%s", ServerBin, value);
+    else
+#endif /* SUPPORT_SNAPPED_CUPSD */
+      cupsdSetEnvf("PATH", "%s/filter:" CUPS_BINDIR ":" CUPS_SBINDIR
+		   ":/bin:/usr/bin", ServerBin);
+  }
   set_if_undefined("SERVER_ADMIN", ServerAdmin);
   set_if_undefined("SHLIB_PATH", NULL);
   set_if_undefined("SOFTWARE", CUPS_MINIMAL);